support float COMPU-NUMERATOR

+++ odxtools/compumethods/readcompumethod.py
@@ -48,7 +48,7 @@ def _parse_compu_scale_to_linear_compu_method(scale_element,
     offset = computation_python_type(next(nums).text)
     factor = computation_python_type(next(nums).text)
     if coeffs.find("COMPU-DENOMINATOR/V") is not None:
-        kwargs["denominator"] = int(
+        kwargs["denominator"] = float(
             coeffs.find("COMPU-DENOMINATOR/V").text)
         assert kwargs["denominator"] > 0
